requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
SQLAlchemy==2.0.37
asyncpg==0.30.0
psycopg2==2.9.10
pydantic==2.10.6
pydantic-settings==2.7.1
passlib[bcrypt]
python-jose==3.3.0
pyjwt==2.10.1
alembic==1.14.1
python-multipart==0.0.20
pytest==8.3.4
pytest-asyncio==0.25.3
httpx==0.28.1
types-passlib
